Removed some lazy loaded queries; fixed up negative stat issues.

+++ dotnetcore/DBComm/Structures/Player.cs
@@ -102,7 +102,7 @@ namespace InfServer.Data
 		{
 			get
 			{
-				return killPoints + assistPoints + bonusPoints;
+				return (long)killPoints + assistPoints + bonusPoints;
 			}
 		}
